# apps/api/tests/fixtures/openapi/api-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "BrandRef": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "BrandRef",
        "type": "object"
      },
...
      "CatalogFacets": {
        "properties": {
          "brands": {
            "items": {
              "$ref": "#/components/schemas/BrandRef"
            },
            "title": "Brands",
            "type": "array"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/app__schemas__v2__catalog__SeriesRef"
            },
            "title": "Series",
            "type": "array"
          },
          "stockStatuses": {
            "items": {
              "enum": [
                "IN_STOCK",
                "PREORDER"
              ],
              "type": "string"
            },
            "title": "Stockstatuses",
            "type": "array"
          }
        },
        "title": "CatalogFacets",
        "type": "object"
      },
      "CatalogProduct": {
        "properties": {
          "attributes": {
            "title": "Attributes",
            "type": "object"
          },
          "basePrice": {
            "$ref": "#/components/schemas/Money"
          },
          "brand": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BrandRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "clientPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/ProductDocument"
            },
            "title": "Documents",
            "type": "array"
          },
          "exchangeRate": {
            "$ref": "#/components/schemas/Rate"
          },
          "hasDiscount": {
            "title": "Hasdiscount",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaResource"
            },
            "title": "Media",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "retailPrice": {
            "$ref": "#/components/schemas/Money"
          },
          "series": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/app__schemas__v2__catalog__SeriesRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "stockQuantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockquantity"
          },
          "stockStatus": {
            "enum": [
              "IN_STOCK",
              "PREORDER"
            ],
            "title": "Stockstatus",
            "type": "string"
          },
          "thumbnail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MediaResource"
              },
              {
                "type": "null"
              }
            ]
          },
          "volumeTiers": {
            "items": {
              "$ref": "#/components/schemas/VolumeTierHint"
            },
            "title": "Volumetiers",
            "type": "array"
          }
        },
        "required": [
          "id",
          "sku",
          "name",
          "stockStatus",
          "basePrice",
          "retailPrice",
          "clientPrice",
          "exchangeRate",
          "hasDiscount"
        ],
        "title": "CatalogProduct",
        "type": "object"
      },
...
      "CursorResponse_CatalogProduct_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CatalogProduct"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/CursorMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "CursorResponse[CatalogProduct]",
        "type": "object"
      },
...
      "ProblemDetails": {
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "minLength": 1,
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ProblemFieldError"
            },
            "title": "Errors",
            "type": "array"
          },
          "instance": {
            "minLength": 1,
            "title": "Instance",
            "type": "string"
          },
          "requestId": {
            "minLength": 1,
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "minLength": 1,
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "requestId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemFieldError": {
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "field": {
            "minLength": 1,
            "title": "Field",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "ProblemFieldError",
        "type": "object"
      },
...
      "ResponseMeta": {
        "properties": {
          "requestId": {
            "examples": [
              "0198f0d2-58f2-79a7-8123-2c72c1548f90"
            ],
            "minLength": 1,
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SuccessResponse_CatalogFacets_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CatalogFacets"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[CatalogFacets]",
        "type": "object"
      },
      "SuccessResponse_CatalogProduct_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CatalogProduct"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SuccessResponse[CatalogProduct]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Клиентский портал с динамическими прайс-листами. Каноничная спека — ARCHITECTURE_PLAN.md / SITEMAP.md.",
    "title": "B2B Price Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/catalog/facets": {
      "get": {
        "operationId": "get_facets_api_v2_catalog_facets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CatalogFacets_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "Get Facets",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v2/catalog/products": {
      "get": {
        "operationId": "list_products_api_v2_catalog_products_get",
        "parameters": [
          {
            "description": "Search by SKU or name",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search by SKU or name",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "brands",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Brands"
            }
          },
          {
            "in": "query",
            "name": "series",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "format": "uuid",
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Series"
            }
          },
          {
            "in": "query",
            "name": "stock",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "IN_STOCK",
                    "PREORDER"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stock"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name",
              "enum": [
                "name",
                "-name",
                "price",
                "-price",
                "sku"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "price_calc_mode",
            "required": false,
            "schema": {
              "default": "fixed",
              "enum": [
                "fixed",
                "nbrb_current"
              ],
              "title": "Price Calc Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorResponse_CatalogProduct_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "List Products",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v2/catalog/products/by-sku/{sku}": {
      "get": {
        "operationId": "get_product_by_sku_api_v2_catalog_products_by_sku__sku__get",
        "parameters": [
          {
            "in": "path",
            "name": "sku",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "title": "Sku",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "price_calc_mode",
            "required": false,
            "schema": {
              "default": "fixed",
              "enum": [
                "fixed",
                "nbrb_current"
              ],
              "title": "Price Calc Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CatalogProduct_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Product not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get Product By Sku",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v2/catalog/products/{productId}": {
      "get": {
        "operationId": "get_product_api_v2_catalog_products__productId__get",
        "parameters": [
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Productid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "price_calc_mode",
            "required": false,
            "schema": {
              "default": "fixed",
              "enum": [
                "fixed",
                "nbrb_current"
              ],
              "title": "Price Calc Mode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_CatalogProduct_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Product not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Get Product",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}